refactor: 공통속성 추상클래스로 정의 및 상속
