+++ src/main/java/com/sprint/mission/discodeit/entity/BinaryContent.java
@@ -1,29 +1,22 @@
 package com.sprint.mission.discodeit.entity;
 
+import com.sprint.mission.discodeit.entity.base.BaseEntity;
 import java.io.Serializable;
-import java.time.Instant;
-import java.util.UUID;
-import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 
 @Getter
 @Builder
-@AllArgsConstructor
-public class BinaryContent implements Serializable {
+public class BinaryContent extends BaseEntity implements Serializable {
 
   private static final long serialVersionUID = 1L;
-  private UUID id;
-  private Instant createdAt;
-  //
   private String fileName;
   private Long size;
   private String contentType;
   private byte[] bytes;
 
   public BinaryContent(String fileName, Long size, String contentType, byte[] bytes) {
-    this.id = UUID.randomUUID();
-    this.createdAt = Instant.now();
+    super();
     this.fileName = fileName;
     this.size = size;
     this.contentType = contentType;
